fix: Review-Findings Systemfirma (Typ-Safety, JPA-Mapping, reload public)
- SystemCompanyEntry.rank: int -> short (konsistent mit SMALLINT/jOOQ Short)
- BusinessPartner: system_rank als read-only JPA-Feld (insertable=false, updatable=false)
- SystemCompanyService.reload(): public, Log nutzt getPrimary() statt cache.get(0)

+++ src/main/java/de/market/businesspartner/model/SystemCompanyEntry.java
@@ -1,3 +1,3 @@
 package de.market.businesspartner.model;
 
-public record SystemCompanyEntry(Long partnerId, String shortName, String name, int rank) {}
+public record SystemCompanyEntry(Long partnerId, String shortName, String name, short rank) {}
